move `materialEnvIntensity` and `materialEnvRotation` to `MaterialProperties`
