+++ src/nodes/accessors/CubeTextureNode.js
@@ -3,7 +3,7 @@ import { reflectVector, refractVector } from './ReflectVector.js';
 import { nodeProxy, vec3 } from '../tsl/TSLBase.js';
 
 import { CubeReflectionMapping, CubeRefractionMapping, WebGPUCoordinateSystem } from '../../constants.js';
-import { materialEnvRotation } from './MaterialNode.js';
+import { materialEnvRotation } from './MaterialProperties.js';
 
 /**
  * This type of uniform node represents a cube texture.
